+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2.7.0] - TBD
 #### Feature
+- Support for Tournament Presets
 - Added commands for recalculating statistics, resetting elo etc
 - New endpoint for getting player hits
 - Support for `ANY` and `MASTER` out for `x01` legs